backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
sqlalchemy==2.0.41
psycopg[binary]==3.2.9
pydantic-settings==2.10.1
celery==5.5.3
redis==6.2.0
python-multipart==0.0.20
qdrant-client==1.14.2
httpx==0.28.1
python-docx==1.1.2
markdown==3.8.2
reportlab==4.4.2
sse-starlette==2.3.6
orjson==3.11.1
yt-dlp==2026.3.17
